Split of effort sharing regimes to module, and ability to pay added
