+++ mimosa/inputdata/config/config_default.yaml
@@ -361,6 +361,7 @@ effort sharing:
       - equal_mitigation_costs
       - equal_total_costs
       - per_cap_convergence
+      - ability_to_pay
     default: noregime
   percapconv_year:
     descr: >-
